Enable build on old Linux (dotnet/coreclr#11414)
This change enables building coreclr on old Linux distros that
don't have PR_SET_PTRACER and don't have _xstate struct in the
standard sigcontext.h header.

Commit migrated from https://github.com/dotnet/coreclr/commit/3437a820fdc94caa3d1775bcee802b056f3adce2

+++ src/clr/src/debug/createdump/threadinfo.cpp
@@ -38,7 +38,7 @@ ThreadInfo::Initialize(ICLRDataTarget* dataTarget)
             return false;
         }
     }
-    TRACE("Thread %04x RIP %016llx RSP %016llx\n", m_tid, m_gpRegisters.rip, m_gpRegisters.rsp);
+    TRACE("Thread %04x RIP %016llx RSP %016llx\n", m_tid, (unsigned long long)m_gpRegisters.rip, (unsigned long long)m_gpRegisters.rsp);
     return true;
 }
